articles/requirements.txt
python-dotenv==0.20.0
asyncpg==0.27.0

# GRPC
grpcio==1.54.0
grpcio-tools==1.54.0